+++ src/features/blog/blogPost/__tests__/newsletter/Newsletter.test.tsx
@@ -1,16 +1,18 @@
-import { render, screen } from '@testing-library/react';
+import { screen } from '@testing-library/react';
+
+import { renderWithQueryClient } from '@utils/tests/queryClient';
 
 import Newsletter from '../../newsletter';
 
 describe('Newsletter component', () => {
   test('renders heading', () => {
-    render(<Newsletter />);
+    renderWithQueryClient(<Newsletter />);
 
     expect(screen.getByRole('heading', { name: /join my newsletter/i })).toBeInTheDocument();
   });
 
   test('renders subscribe form', () => {
-    render(<Newsletter />);
+    renderWithQueryClient(<Newsletter />);
 
     expect(screen.getByRole('form', { name: /newsletter subscribe/i })).toBeInTheDocument();
   });